edición video 9

+++ app/Livewire/Instructor/Courses/ManageLessons.php
@@ -17,7 +17,6 @@ class ManageLessons extends Component
     public $lessons;
     public $video;
     public $url;
-    public $document;
 
     public $lessonCreate = [
         'open' => false,
@@ -41,8 +40,6 @@ class ManageLessons extends Component
         'existing_video_path' => null,
     ];
 
-    public $orderLessons;
-
     protected $listeners = ['refreshOrderLessons' => 'getLessons'];
 
     public function mount()
@@ -63,15 +60,17 @@ class ManageLessons extends Component
             'lessonCreate.name' => ['required', new UniqueLessonCourse($this->section->course_id)],
             'lessonCreate.platform' => 'required|in:1,2',
             'lessonCreate.description' => 'nullable|string',
-            'lessonCreate.document' => 'nullable|mimes:pdf|max:2048',
+            'lessonCreate.document' => 'nullable|file|mimes:pdf|max:2048',
         ];
     }
 
-    public function store()
+    // Método para almacenar nueva lección
+    public function store() 
     {
+        // Validamos las entradas
         $this->validate();
 
-        // Recopilar datos para la lección
+        // Acción para almacenar
         $lessonData = [
             'name' => $this->lessonCreate['name'],
             'description' => $this->lessonCreate['description'],
@@ -80,43 +79,45 @@ class ManageLessons extends Component
         ];
 
         try {
-            // Manejo del upload del documento PDF
-            if ($this->lessonCreate['document'] instanceof UploadedFile) {
+            // Procesa el documento solamente si existe
+            if (isset($this->lessonCreate['document']) && $this->lessonCreate['document'] instanceof UploadedFile) {
+                // Almacena el documento
                 $lessonData['document_path'] = $this->lessonCreate['document']->store('courses/documents', 'public');
                 $lessonData['document_original_name'] = $this->lessonCreate['document']->getClientOriginalName();
-            }
+            } 
 
-            // Manejo del upload del video
-            if ($lessonData['platform'] == 1 && $this->video instanceof UploadedFile) {
+            // Manejo del video
+            if ($lessonData['platform'] == 1 && isset($this->video) && $this->video instanceof UploadedFile) {
+                // Almacena el video
                 $lessonData['video_path'] = $this->video->store('courses/lessons', 'public');
                 $lessonData['video_original_name'] = $this->video->getClientOriginalName();
             } elseif ($lessonData['platform'] == 2) {
-                $lessonData['video_path'] = null;
+                // Para YouTube
+                $lessonData['video_path'] = null; 
                 $lessonData['video_original_name'] = $this->url;
             }
 
-            // Crear la nueva lección en la base de datos
+            // Creación en la base de datos
             Lesson::create($lessonData);
-
-            // Reiniciar los valores para la creación de la lección
-            $this->reset(['url', 'lessonCreate', 'video', 'document']);
-            $this->getLessons();
+            $this->reset(['url', 'lessonCreate', 'video']);
+            $this->getLessons(); // Refresca la lista
             $this->emit('refreshOrderLessons');
-
         } catch (\Exception $e) {
-            // Manejo de errores, captura errores y muestra el mensaje
+            // Captura el error
             $this->dispatchBrowserEvent('notify', ['message' => 'Error: ' . $e->getMessage()]);
         }
     }
 
-    public function edit($lessonId)
+    // Método de edición
+    public function edit($lessonId) 
     {
         $lesson = Lesson::findOrFail($lessonId);
+        // Asigna los valores actuales
         $this->lessonEdit = [
             'id' => $lesson->id,
             'name' => $lesson->name,
             'description' => $lesson->description,
-            'document' => null,
+            'document' => $lesson->document_path ? Storage::url($lesson->document_path) : null, // Mantener la URL
             'document_path' => $lesson->document_path,
             'document_original_name' => $lesson->document_original_name,
             'platform' => $lesson->platform,
@@ -125,8 +126,10 @@ class ManageLessons extends Component
         ];
     }
 
-    public function update()
+    // Método de actualización
+    public function update() 
     {
+        // Validación
         $this->validate([
             'lessonEdit.name' => ['required'],
             'lessonEdit.description' => ['nullable'],
@@ -135,84 +138,78 @@ class ManageLessons extends Component
 
         $lesson = Lesson::findOrFail($this->lessonEdit['id']);
 
-        // Comienza la actualización de la lección
         try {
+            // Actualiza los datos
             $lesson->update([
                 'name' => $this->lessonEdit['name'],
                 'description' => $this->lessonEdit['description'],
             ]);
 
-            // Manejo del archivo del documento
-            if ($this->lessonEdit['document'] instanceof UploadedFile) {
-                // Eliminar el documento viejo
+            // Solo intentar guardar el documento si uno nuevo ha sido cargado
+            if (isset($this->lessonEdit['document']) && $this->lessonEdit['document'] instanceof UploadedFile) {
+                // Elimina el documento viejo
                 if ($lesson->document_path && Storage::exists($lesson->document_path)) {
                     Storage::delete($lesson->document_path);
                 }
 
-                // Guardar el nuevo documento
+                // Almacena nuevo documento
                 $lesson->document_path = $this->lessonEdit['document']->store('courses/documents', 'public');
                 $lesson->document_original_name = $this->lessonEdit['document']->getClientOriginalName();
-                $lesson->save();
             }
 
             // Manejo del video
-            if ($lesson->platform == 1 && $this->lessonEdit['video'] instanceof UploadedFile) {
+            if ($lesson->platform == 1 && isset($this->lessonEdit['video']) && $this->lessonEdit['video'] instanceof UploadedFile) {
+                // Elimina el video viejo si hay uno
                 if ($lesson->video_path && Storage::exists($lesson->video_path)) {
                     Storage::delete($lesson->video_path);
                 }
+                // Almacena nuevo video
                 $lesson->video_path = $this->lessonEdit['video']->store('courses/lessons', 'public');
                 $lesson->video_original_name = $this->lessonEdit['video']->getClientOriginalName();
-                $lesson->save();
             } elseif ($lesson->platform == 2) {
-                // No se guarda video si es YouTube
-                if ($lesson->video_path && Storage::exists($lesson->video_path)) {
-                    Storage::delete($lesson->video_path);
-                }
-                $lesson->video_path = null;
+                // Si es vídeo de YouTube
+                $lesson->video_path = null; 
                 $lesson->video_original_name = $this->lessonEdit['url'];
-                $lesson->save();
             }
 
+            $lesson->save();
             $this->reset('lessonEdit');
             $this->getLessons();
-
         } catch (\Exception $e) {
-            // Manejo de errores
             $this->dispatchBrowserEvent('notify', ['message' => 'Error: ' . $e->getMessage()]);
         }
     }
 
-    public function sortLessons($order)
-    {
-        foreach ($order as $index => $lessonId) {
-            Lesson::findOrFail($lessonId)->update(['position' => $index + 1]);
-        }
-        $this->getLessons();
-        $this->emit('refreshOrderLessons');
-    }
-
-    public function destroy($lessonId)
+    // Método de eliminación
+    public function destroy($lessonId) 
     {
         $lesson = Lesson::findOrFail($lessonId);
 
+        // Eliminar el video relacionado
         if ($lesson->video_path && Storage::exists($lesson->video_path)) {
             Storage::delete($lesson->video_path);
         }
 
+        // Eliminar el documento relacionado
         if ($lesson->document_path && Storage::exists($lesson->document_path)) {
             Storage::delete($lesson->document_path);
         }
 
+        // Eliminar la lección en sí
         $lesson->delete();
         $this->getLessons();
-        $this->emit('refreshOrderLessons');
+        $this->emit('refreshOrderLessons'); // Refrescar la lista
     }
 
-    public function render()
+    // Rendirizar la vista
+    public function render() 
     {
         return view('livewire.instructor.courses.manage-lessons');
     }
 }
+
+
+
